change name hooks of shapes and refactor elements  mappers
