+++ src/editor/core/hooks/index.ts
@@ -1,6 +1,5 @@
-export { default as useSelectedShape } from "./element/hook";
+export { default as useSelectedShape } from "./selectedShape/hook";
 export { default as useEvent } from "./events/hook";
-export { default as useSelection } from "./selection/hook";
 export { default as useStage } from "./stage/hook";
 export { default as useTool } from "./tool/hook";
 export { default as useZoom } from "./zoom/hook";
