--- src/editor/core/hooks/selection/jotai.ts
@@ -1,16 +1,0 @@
-import { atom } from "jotai";
-import Konva from "konva";
-import { MutableRefObject } from "react";
-
-type ISelectionRefs = {
-  rectRef: MutableRefObject<any>;
-  trRef: MutableRefObject<any>;
-  selection: MutableRefObject<any>;
-  layerRef: MutableRefObject<Konva.Layer | undefined>;
-};
-
-const selectionAtom = atom({} as ISelectionRefs);
-
-export const isSelectElementsAtom = atom(false);
-
-export default selectionAtom;
