--- src/editor/core/hooks/useStroke/hook.ts
@@ -1,10 +1,0 @@
-import { useAtomValue } from "jotai";
-import strokeAtom from "./jotai";
-
-const useStroke = () => {
-  const stroke = useAtomValue(strokeAtom);
-
-  return stroke;
-};
-
-export default useStroke;
