+++ src/pages/_document.tsx
@@ -3,17 +3,7 @@ import { Head, Html, Main, NextScript } from "next/document";
 export default function Document() {
   return (
     <Html lang="en">
-      <Head>
-        <meta
-          name="viewport"
-          content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"
-        />
-        <script
-          defer
-          data-domain="harmony.whil.online"
-          src="https://plausible.io/js/script.js"
-        ></script>
-      </Head>
+      <Head />
       <body>
         <Main />
         <NextScript />
